server/requirements.txt
coverage==4.2
defusedxml==0.4.1
Django==1.10.4
django-braces==1.10.0
django-cors-headers==1.3.1
django-crequest==2014.9.19
django-dbbackup==3.1.3
django-method-override==0.2.0
django-oauth-toolkit==0.11.0
django-post-office==2.0.8
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.5.3
jsonfield==1.0.3
oauthlib==1.1.2
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
PyJWT==1.4.2
python-magic==0.4.12
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2016.10
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
stripe==1.46.0
